fix(client): hydration was not working at all for example
Some concepts were changed with this PR:

* As inline scripts cannot be defered we exchanged the JS with JSON
string. This allows to render them close to the chunks but without
interrupting the browser.
* The runtime is becomming way smaller to reduce js in blocking inline script
* Only the entry js bundle is added to the html. The chunks are still
prefetched using the link header. This has the benefit that we can
execute the main bundle async, so before the dom has been parsed. This should increase the startup time as we can also fetch and execute the chunk bundles while the HTML is generated.