fix(tests): glorious return of the babel config
The config was actually required to run Jest.
It was removed because of some errors when compiling after dependency
updates.
It was actually because of a super modern target, therefore less code
was transpiled. Which caused follow up issues (not sure if in parcel or
somewhere else).
A lower target slowes it by now, hope further updates od deps fix it...