feat(plugin): implement styled components plugin
This introduces an interface for a plugin system. The current
implementation does not make too much sense but allows to be extensible
in the future to implement different kind of plugins.

It would most probably make most sense to wait for another kind of
plugin to see which abstraction does make sense.