fix(Chunk): invalidate vie state if hydrated from cache
When the chunk was hydrated from cache, e.g. at mount,
it was not aware of the used resourced and therefore thought
it is always valid.

To solve this all used resource ids are persisted in the view state.
Therfore it was required to implement resource ids which are shareable
between the server and the client.

The ids are not perfect and can overlap in some unfortuned situations.
We should add warnings for this cases and maybe offer the user to set a
key or similar.

Further the types for the resource creator are even more broken
now, as the name is optional for the user but required internally.
The requests types and file needs a refactor anyhow and should therefor
be done within one of the next PRs.