fix(server): sync correct resource ids to client
We cloned the request twice per chunk as we created one
for the view state cache and one  for each request.

Therefore the wrong resource ids were retrieved and injected
into the view.

As the view state cache injection was a pain before
and didn't made sense any longer it was merged with the
requestDataForChunks function. Therefore we now have one
function that can do all the data fetching on its own.