feat(core): push resources from server after html is rendered
Resources are expected to be reused on the client. If they are not
pushed from the server the client would need to fetch them all again.
But those resources are not required for the initial rendering as the
view state is already pushed.
Therefore we first render all html and tell the client to hydrate, and
push the remaining resources lazy at the end.
Those resources will be delivered as a pure string, therefore they will
only be evaluated as soon as they are requested from the cache.

Closes #24