feat(chunk): implement concurrent mode
Adds an experimental feature flag PEXO_EXPERIMENTAL.
If set to `true` it will run the application in concurrent mode.
Currently this is only makes changes to the client, the server is not
affected so far.

There is some more potential to make use of concurrent mode for deferred
values and priorization of components. This might follow as a next step.

This commit contains a breaking change as it renames `Loading` and
`Error` to `LoadingView` and `ErrorView` as the former API for the error
component caused unintended in case `throw new Error()` was used.

Closes #10